refactor: rename de dto path names
