+++ employee/src/main/java/com/jobby/employee/infraestructure/adapters/in/rest/dto/created/AddressCreated.java
@@ -1,4 +1,4 @@
-package com.jobby.employee.infraestructure.dtos.objects.created;
+package com.jobby.employee.infraestructure.adapters.in.rest.dto.created;
 
 import lombok.Getter;
 import lombok.Setter;
